% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOut.r
\name{filterOut}
\alias{filterOut}
\title{Filter routine}
\usage{
filterOut(datasett, filter1, filter2, filter3)
}
\arguments{
\item{data}{}

\item{filter}{}
}
\value{
datasett
}
\description{
Filter routine
}
