% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.r
\name{makeGeojsonMap}
\alias{makeGeojsonMap}
\title{Title}
\usage{
makeGeojsonMap(data = NULL, map = NULL)
}
\arguments{
\item{data}{}

\item{map}{}
}
\value{
a map made from a geojson file
}
\description{
Title
}
