% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_module.r
\name{map_server}
\alias{map_server}
\title{Function to make a map}
\usage{
map_server(id, data = NULL, map = NULL)
}
\arguments{
\item{id}{}

\item{data}{Data to be plotted in the map}

\item{map}{The map itself}
}
\value{
a map
}
\description{
Function to make a map
}
