% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readJS.R
\name{plotVariasjon}
\alias{plotVariasjon}
\title{Make a plot based on a data frame}
\usage{
plotVariasjon(fullDataFrame = NULL, which_id = "i0")
}
\arguments{
\item{fullDataFrame}{Data to be plotted}

\item{which_id}{Which data to plot (with id == which_id)}
}
\description{
Make a plot based on a data frame
}
