% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.r
\name{makeGeojsonMap}
\alias{makeGeojsonMap}
\title{Plot a map with data}
\usage{
makeGeojsonMap(data = NULL, map = NULL)
}
\arguments{
\item{data}{Data to be plotted}

\item{map}{Map to plot}
}
\value{
a plot made from a geojson map
}
\description{
Plot a map with data
}
